tkt frérot

+++ src/machine.h
@@ -1,5 +1,9 @@
 #pragma once
 
+#include <stdlib.h>
+#include <stdbool.h>
+#include <string.h>
+
 typedef enum MachineType{
     STATION=1,
     SWITCH=2
@@ -8,7 +12,7 @@ typedef enum MachineType{
 typedef struct machine
 {
     enum MachineType type;
-    int numMachine;
+    size_t numMachine;
     void* ptr;
 }machine;
